Extract `FirebaseAppDistributionStub`.
