+++ firebase-appdistribution-api/src/main/java/com/google/firebase/appdistribution/internal/FirebaseAppDistributionProxy.java
@@ -38,254 +38,49 @@ import java.util.concurrent.Executor;
  * This implementation of the Firebase App Distribution API proxies to the real implementation
  * ({@code FirebaseAppDistributionImpl}) provided by the {@code
  * com.google.firebase:firebase-appdistribution} artifact. If that artifact is not included in the
- * build then the stubs will return failed {@link Task Tasks}/{@link UpdateTask UpdateTasks} with
+ * build, then the stubs will return failed {@link Task Tasks}/{@link UpdateTask UpdateTasks} with
  * {@link FirebaseAppDistributionException.Status#NOT_IMPLEMENTED}.
  */
 public class FirebaseAppDistributionProxy implements FirebaseAppDistribution {
-  private final Provider<FirebaseAppDistribution> firebaseAppDistributionImplProvider;
+  private final FirebaseAppDistribution delegate;
 
   public FirebaseAppDistributionProxy(
       Provider<FirebaseAppDistribution> firebaseAppDistributionImplProvider) {
-    this.firebaseAppDistributionImplProvider = firebaseAppDistributionImplProvider;
+    FirebaseAppDistribution impl = firebaseAppDistributionImplProvider.get();
+    delegate = impl != null ? impl : new FirebaseAppDistributionStub();
   }
 
   @NonNull
   @Override
   public UpdateTask updateIfNewReleaseAvailable() {
-    if (firebaseAppDistributionImplProvider.get() == null) {
-      return new NotImplementedUpdateTask();
-    }
-    return firebaseAppDistributionImplProvider.get().updateIfNewReleaseAvailable();
+    return delegate.updateIfNewReleaseAvailable();
   }
 
   @Override
   public boolean isTesterSignedIn() {
-    if (firebaseAppDistributionImplProvider.get() == null) {
-      return false;
-    }
-    return firebaseAppDistributionImplProvider.get().isTesterSignedIn();
+    return delegate.isTesterSignedIn();
   }
 
   @NonNull
   @Override
   public Task<Void> signInTester() {
-    if (firebaseAppDistributionImplProvider.get() == null) {
-      return getNotImplementedTask();
-    }
-    return firebaseAppDistributionImplProvider.get().signInTester();
+    return delegate.signInTester();
   }
 
   @Override
   public void signOutTester() {
-    if (firebaseAppDistributionImplProvider.get() == null) {
-      return;
-    }
-    firebaseAppDistributionImplProvider.get().signOutTester();
+    delegate.signOutTester();
   }
 
   @NonNull
   @Override
   public synchronized Task<AppDistributionRelease> checkForNewRelease() {
-    if (firebaseAppDistributionImplProvider.get() == null) {
-      return getNotImplementedTask();
-    }
-    return firebaseAppDistributionImplProvider.get().checkForNewRelease();
+    return delegate.checkForNewRelease();
   }
 
   @NonNull
   @Override
   public UpdateTask updateApp() {
-    if (firebaseAppDistributionImplProvider.get() == null) {
-      return new NotImplementedUpdateTask();
-    }
-    return firebaseAppDistributionImplProvider.get().updateApp();
-  }
-
-  private static <TResult> Task<TResult> getNotImplementedTask() {
-    return Tasks.forException(
-        new FirebaseAppDistributionException(
-            "This API is not implemented. The build was compiled against the API only.",
-            Status.NOT_IMPLEMENTED));
-  }
-
-  private static class NotImplementedUpdateTask extends UpdateTask {
-    private final Task<Void> task = getNotImplementedTask();
-
-    @NonNull
-    @Override
-    public <TContinuationResult> Task<TContinuationResult> continueWith(
-        @NonNull Continuation<Void, TContinuationResult> continuation) {
-      return task.continueWith(continuation);
-    }
-
-    @NonNull
-    @Override
-    public <TContinuationResult> Task<TContinuationResult> continueWith(
-        @NonNull Executor executor, @NonNull Continuation<Void, TContinuationResult> continuation) {
-      return task.continueWith(executor, continuation);
-    }
-
-    @NonNull
-    @Override
-    public <TContinuationResult> Task<TContinuationResult> continueWithTask(
-        @NonNull Continuation<Void, Task<TContinuationResult>> continuation) {
-      return task.continueWithTask(continuation);
-    }
-
-    @NonNull
-    @Override
-    public <TContinuationResult> Task<TContinuationResult> continueWithTask(
-        @NonNull Executor executor,
-        @NonNull Continuation<Void, Task<TContinuationResult>> continuation) {
-      return task.continueWithTask(executor, continuation);
-    }
-
-    @NonNull
-    @Override
-    public <TContinuationResult> Task<TContinuationResult> onSuccessTask(
-        @NonNull SuccessContinuation<Void, TContinuationResult> successContinuation) {
-      return task.onSuccessTask(successContinuation);
-    }
-
-    @NonNull
-    @Override
-    public <TContinuationResult> Task<TContinuationResult> onSuccessTask(
-        @NonNull Executor executor,
-        @NonNull SuccessContinuation<Void, TContinuationResult> successContinuation) {
-      return task.onSuccessTask(executor, successContinuation);
-    }
-
-    @NonNull
-    @Override
-    public UpdateTask addOnProgressListener(@NonNull OnProgressListener listener) {
-      return this;
-    }
-
-    @NonNull
-    @Override
-    public UpdateTask addOnProgressListener(
-        @Nullable Executor executor, @NonNull OnProgressListener listener) {
-      return this;
-    }
-
-    @Override
-    public boolean isComplete() {
-      return task.isComplete();
-    }
-
-    @Override
-    public boolean isSuccessful() {
-      return task.isSuccessful();
-    }
-
-    @Override
-    public boolean isCanceled() {
-      return task.isCanceled();
-    }
-
-    @Nullable
-    @Override
-    public Void getResult() {
-      return task.getResult();
-    }
-
-    @Nullable
-    @Override
-    public <X extends Throwable> Void getResult(@NonNull Class<X> aClass) throws X {
-      return task.getResult(aClass);
-    }
-
-    @Nullable
-    @Override
-    public Exception getException() {
-      return task.getException();
-    }
-
-    @NonNull
-    @Override
-    public Task<Void> addOnSuccessListener(
-        @NonNull OnSuccessListener<? super Void> onSuccessListener) {
-      task.addOnSuccessListener(onSuccessListener);
-      return this;
-    }
-
-    @NonNull
-    @Override
-    public Task<Void> addOnSuccessListener(
-        @NonNull Executor executor, @NonNull OnSuccessListener<? super Void> onSuccessListener) {
-      task.addOnSuccessListener(executor, onSuccessListener);
-      return this;
-    }
-
-    @NonNull
-    @Override
-    public Task<Void> addOnSuccessListener(
-        @NonNull Activity activity, @NonNull OnSuccessListener<? super Void> onSuccessListener) {
-      task.addOnSuccessListener(activity, onSuccessListener);
-      return this;
-    }
-
-    @NonNull
-    @Override
-    public Task<Void> addOnFailureListener(@NonNull OnFailureListener onFailureListener) {
-      task.addOnFailureListener(onFailureListener);
-      return this;
-    }
-
-    @NonNull
-    @Override
-    public Task<Void> addOnFailureListener(
-        @NonNull Executor executor, @NonNull OnFailureListener onFailureListener) {
-      task.addOnFailureListener(executor, onFailureListener);
-      return this;
-    }
-
-    @NonNull
-    @Override
-    public Task<Void> addOnFailureListener(
-        @NonNull Activity activity, @NonNull OnFailureListener onFailureListener) {
-      task.addOnFailureListener(activity, onFailureListener);
-      return this;
-    }
-
-    @NonNull
-    @Override
-    public Task<Void> addOnCompleteListener(@NonNull OnCompleteListener<Void> onCompleteListener) {
-      return task.addOnCompleteListener(onCompleteListener);
-    }
-
-    @NonNull
-    @Override
-    public Task<Void> addOnCompleteListener(
-        @NonNull Executor executor, @NonNull OnCompleteListener<Void> onCompleteListener) {
-      return task.addOnCompleteListener(executor, onCompleteListener);
-    }
-
-    @NonNull
-    @Override
-    public Task<Void> addOnCompleteListener(
-        @NonNull Activity activity, @NonNull OnCompleteListener<Void> onCompleteListener) {
-      return task.addOnCompleteListener(activity, onCompleteListener);
-    }
-
-    @NonNull
-    @Override
-    public Task<Void> addOnCanceledListener(@NonNull OnCanceledListener onCanceledListener) {
-      return task.addOnCanceledListener(onCanceledListener);
-    }
-
-    @NonNull
-    @Override
-    public Task<Void> addOnCanceledListener(
-        @NonNull Executor executor, @NonNull OnCanceledListener onCanceledListener) {
-      return task.addOnCanceledListener(executor, onCanceledListener);
-    }
-
-    @NonNull
-    @Override
-    public Task<Void> addOnCanceledListener(
-        @NonNull Activity activity, @NonNull OnCanceledListener onCanceledListener) {
-      return task.addOnCanceledListener(activity, onCanceledListener);
-    }
+    return delegate.updateApp();
   }
 }
